apps/agent-service/pyproject.toml
[project]
name = "evento-agent-service"
version = "0.1.0"
description = "Multi-agent event content creation service using Google ADK"
requires-python = ">=3.14"
dependencies = [
    "google-adk>=0.6.0",
]

